+++ packages/node/src/declarations.d.ts
@@ -1,2 +1,3 @@
 declare module 'lsmod';
 declare module 'https-proxy-agent';
+declare module 'async-limiter';