+++ packages/node/test/helper/settimeouttransport.ts
@@ -1,14 +1,16 @@
-import { SentryEvent, SentryResponse, Status } from '../../src';
+import { SentryResponse, Status } from '../../src';
 import { BaseTransport } from '../../src/transports';
 
 export class SetTimeoutTransport extends BaseTransport {
-  public async captureEvent(_: SentryEvent): Promise<SentryResponse> {
-    return new Promise<SentryResponse>(resolve => {
-      setTimeout(() => {
-        resolve({
-          status: Status.fromHttpCode(200),
-        });
-      }, 1);
-    });
+  public async sendEvent(_: string): Promise<SentryResponse> {
+    return this.buffer.add(
+      new Promise<SentryResponse>(resolve => {
+        setTimeout(() => {
+          resolve({
+            status: Status.fromHttpCode(200),
+          });
+        }, 1);
+      }),
+    );
   }
 }