+++ packages/node/test/transports/https.test.ts
@@ -50,9 +50,11 @@ describe('HTTPSTransport', () => {
 
   test('send 200', async () => {
     const transport = createTransport({ dsn });
-    await transport.captureEvent({
-      message: 'test',
-    });
+    await transport.sendEvent(
+      JSON.stringify({
+        message: 'test',
+      }),
+    );
 
     const requestOptions = (transport.module!.request as jest.Mock).mock.calls[0][0];
     assertBasicOptions(requestOptions);
@@ -64,9 +66,11 @@ describe('HTTPSTransport', () => {
     const transport = createTransport({ dsn });
 
     try {
-      await transport.captureEvent({
-        message: 'test',
-      });
+      await transport.sendEvent(
+        JSON.stringify({
+          message: 'test',
+        }),
+      );
     } catch (e) {
       const requestOptions = (transport.module!.request as jest.Mock).mock.calls[0][0];
       assertBasicOptions(requestOptions);
@@ -82,9 +86,11 @@ describe('HTTPSTransport', () => {
     const transport = createTransport({ dsn });
 
     try {
-      await transport.captureEvent({
-        message: 'test',
-      });
+      await transport.sendEvent(
+        JSON.stringify({
+          message: 'test',
+        }),
+      );
     } catch (e) {
       const requestOptions = (transport.module!.request as jest.Mock).mock.calls[0][0];
       assertBasicOptions(requestOptions);
@@ -100,9 +106,11 @@ describe('HTTPSTransport', () => {
         a: 'b',
       },
     });
-    await transport.captureEvent({
-      message: 'test',
-    });
+    await transport.sendEvent(
+      JSON.stringify({
+        message: 'test',
+      }),
+    );
 
     const requestOptions = (transport.module!.request as jest.Mock).mock.calls[0][0];
     assertBasicOptions(requestOptions);
@@ -115,9 +123,11 @@ describe('HTTPSTransport', () => {
       dsn,
       httpsProxy: 'https://example.com:8080',
     });
-    await transport.captureEvent({
-      message: 'test',
-    });
+    await transport.sendEvent(
+      JSON.stringify({
+        message: 'test',
+      }),
+    );
 
     const requestOptions = (transport.module!.request as jest.Mock).mock.calls[0][0];
     assertBasicOptions(requestOptions);
@@ -134,9 +144,11 @@ describe('HTTPSTransport', () => {
       caCerts: './some/path.pem',
       dsn,
     });
-    await transport.captureEvent({
-      message: 'test',
-    });
+    await transport.sendEvent(
+      JSON.stringify({
+        message: 'test',
+      }),
+    );
     const requestOptions = (transport.module!.request as jest.Mock).mock.calls[0][0];
     assertBasicOptions(requestOptions);
     expect(requestOptions.ca).toEqual('mockedCert');